+++ src/config/index.ts
@@ -1,0 +1,6 @@
+export const CORS_CONFIG = {
+  origin: ['http://localhost:3000', 'http://localhost:8080'],
+  methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
+  preflightContinue: false,
+  optionsSuccessStatus: 204,
+};
