+++ tsconfig.json
@@ -34,7 +34,10 @@
       "@/routes/*": ["routes/*"],
       "@/types/*": ["types/*"],
       "@/lib/*": ["lib/*"],
-      "@/database/*": ["database/*"]
+      "@/database/*": ["database/*"],
+      // "@/config/*": ["config/*"],
+      "@/config": ["config/index.ts"],
+      "@/utils/*": ["utils/*"]
     } /* Specify a set of entries that re-map imports to additional lookup locations. */,
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
@@ -51,7 +54,7 @@
 
     /* JavaScript Support */
     // "allowJs": true,                                  /* Allow JavaScript files to be a part of your program. Use the 'checkJS' option to get errors from these files. */
-    // "checkJs": true,                                  /* Enable error reporting in type-checked JavaScript files. */
+    // "checkJs": true /* Enable error reporting in type-checked JavaScript files. */,
     // "maxNodeModuleJsDepth": 1,                        /* Specify the maximum folder depth used for checking JavaScript files from 'node_modules'. Only applicable with 'allowJs'. */
 
     /* Emit */
